// src/app/models/user.model.ts
export class UserModel {
  agency_id: number;
  agency_logo: string;
  agency_name: string;
  Agent: string;
  booking_id: string;
  Customer: string;
  CustomerFirstName: string;
  CustomerLastName: string;
  driver_id: string;
  drop_details: any;
  group_name: string;
  hub_list: any;
  itinerary: any;
  members_data: any;
  order_id: string;
  Phone: string;
  pickup_details: any;
  travel_date: string;
}